--- src/components/home/ClickTopping/index.tsx
@@ -1,15 +1,0 @@
-export const OpenQuiz = () => {
-  return <div>Open Quiz</div>;
-};
-
-export const ReplyOrNot = () => {
-  return <div>Reply or Not</div>;
-};
-
-export const ReplyLetter = () => {
-  return <div>Reply Letter</div>;
-};
-
-export const ReadMessage = () => {
-  return <div>Read Message</div>;
-};
