+++ src/components/owner/OpenQuiz/index.tsx
@@ -1,0 +1,4 @@
+export const OpenQuiz = () => {
+    return <div>Open Quiz</div>;
+  };
+
