+++ src/components/owner/ReadMessage/index.tsx
@@ -1,0 +1,4 @@
+export const ReadMessage = () => {
+    return <div>Read Message</div>;
+  };
+
