+++ src/components/owner/ReplyLetter/index.tsx
@@ -1,0 +1,5 @@
+export const ReplyLetter = () => {
+    return <div>Reply Letter</div>;
+  };
+  
+
