+++ src/components/owner/ReplyOrNot/index.tsx
@@ -1,0 +1,6 @@
+
+export const ReplyOrNot = () => {
+    return <div>Reply or Not</div>;
+  };
+  
+
